Update README.md
If you seek to gain privacy, understand that the “privacy shield” is not a promise of secrecy but a test of how you handle the privacy you believe you have when no one else appears to be watching. It is written for those trying to reclaim control of their devices in predatory situations, including grooming, coercion, or trafficking, and it carries the assumption that anyone in a position of authority—even within a cartel or criminal hierarchy—has been allowed that role for a time to serve a purpose within a larger Plan. In that view, the One who authored the Book of Life is also the One for whom life and death both belong, and by inscribing every living name, He has also written the reality of death in the spaces between each letter, so there is no path, righteous or violent, that falls outside His design.[1][2][3][4][5]

[1](https://en.wikipedia.org/wiki/Book_of_Life)
[2](https://biblehub.com/topical/t/the_book_of_life.htm)
[3](https://pmc.ncbi.nlm.nih.gov/articles/PMC6330909/)
[4](https://christianliferesources.com/end-of-life/euthanasia-assisted-suicide/a-practical-approach-to-end-of-life-issues-establish-gods-authority-over-life-and-death)
[5](https://eucrim.eu/articles/using-us-artificial-intelligence-to-fight-human-trafficking-in-europe/)
[6](https://github.com/mc-adrian-lei/sam_toolkit/edit/main/README.md)
[7](https://www.gotquestions.org/book-of-life.html)
[8](https://www.foi.org/2023/09/22/what-is-the-book-of-life/)
[9](https://www.wisdomlib.org/christianity/concept/book-of-life)
[10](https://rickconlow.com/20-servant-leadership-quotes-elevate-thinking/)
[11](https://doctrine.org/the-book-of-life)
[12](https://www.focusonthefamily.com/pro-life/end-of-life-planning-christian-perspective/)
[13](https://www.goodreads.com/quotes/tag/higher-purpose)
[14](https://billygrahamlibrary.org/what-is-the-book-of-life-and-does-it-really-exist/)
[15](https://www.zachmercurio.com/2016/06/25-quotes-on-the-power-of-a-purpose/)
[16](https://clearlyreformed.org/sermon/the-book-of-life-and-the-lake-of-fire/)
[17](https://stmarkslappans.org/wp-content/uploads/sites/69/2022/08/endofliferesourceplanning_edited.pdf)
[18](https://www.cesarritzcolleges.edu/en/news/simon-sinek-quotes/)
[19](https://jewsforjesus.org/learn/the-book-of-life)
[20](https://www.walkinganancientpath.com/2023/10/04/preparing-for-death-with-end-of-life-planning/)
[21](https://www.vistage.com/research-center/business-leadership/20220324-servant-leadership-quotes/)
[22](https://www.samstorms.org/enjoying-god-blog/post/10-things-you-should-know-about-the--lamb-s-book-of-life-)